+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@ Version 3.42.0-eisop5 (November ??, 2024)
 -----------------------------------------
 
 **User-visible changes:**
+New command-line options:
+* `-AignoreDeadCode`: Ignore unreachable code when running Checkers.
 
 Removed support for the `-Anocheckjdk` option, which was deprecated in version 3.1.1.
 Use `-ApermitMissingJdk` instead.